more general solution to showing SignInOrAgeDialog
